Simplify FeeManager trait
